// src/app/Models/cxp/forwards-model.ts
export class Forwards {
    IdForward: number; 
    FechaCierre: Date;
    FechaPago: Date;
    CantidadDlls: string;
    TipoCambio: string;
    CantidadMXN: string;
    Garantia: string;
    GarantiaPagada: string;
    CantidadPendiente: string;
    Destino: string;
    Promedio: string;
    Estatus: string;
}